sorting and filtering

+++ dev-data/data/import-dev-data.js
@@ -1,60 +1,39 @@
-const fs = require('fs');
-const mongoose = require('mongoose');
-const dotenv = require('dotenv');
-const Tour = require('./../../models/tourModel');
-const Review = require('./../../models/reviewModel');
-const User = require('./../../models/userModel');
-
-dotenv.config({ path: './config.env' });
-
-const DB = process.env.DATABASE.replace(
-  '<PASSWORD>',
-  process.env.DATABASE_PASSWORD
-);
-
-mongoose
-  .connect(DB, {
+let fs = require('fs')
+let Tour = require('../../models/tourModel')
+let mongoose = require('mongoose')
+let dotenv = require('dotenv')
+dotenv.config({path: './config.env'})
+let DB = process.env.DATABASE_LOCAL
+mongoose.connect(DB, {
     useNewUrlParser: true,
     useCreateIndex: true,
-    useFindAndModify: false
-  })
-  .then(() => console.log('DB connection successful!'));
+    useFindAndModify: false,
+    useUnifiedTopology: true
+})
+.then(() => console.log('db connected!'))
 
-// READ JSON FILE
-const tours = JSON.parse(fs.readFileSync(`${__dirname}/tours.json`, 'utf-8'));
-const users = JSON.parse(fs.readFileSync(`${__dirname}/users.json`, 'utf-8'));
-const reviews = JSON.parse(
-  fs.readFileSync(`${__dirname}/reviews.json`, 'utf-8')
-);
-
-// IMPORT DATA INTO DB
-const importData = async () => {
+let tours = JSON.parse(fs.readFileSync(`${__dirname}/tours-simple.json`, 'utf-8'))
+let importData = async () => {
   try {
-    await Tour.create(tours);
-    await User.create(users, { validateBeforeSave: false });
-    await Review.create(reviews);
-    console.log('Data successfully loaded!');
-  } catch (err) {
-    console.log(err);
+    await Tour.create(tours)
+    console.log('data successfuly loaded!')
+  } catch(e) {
+    console.log(e)
   }
-  process.exit();
-};
-
-// DELETE ALL DATA FROM DB
-const deleteData = async () => {
-  try {
-    await Tour.deleteMany();
-    await User.deleteMany();
-    await Review.deleteMany();
-    console.log('Data successfully deleted!');
-  } catch (err) {
-    console.log(err);
-  }
-  process.exit();
-};
-
-if (process.argv[2] === '--import') {
-  importData();
-} else if (process.argv[2] === '--delete') {
-  deleteData();
+  process.exit()
 }
+
+// delete all data from collection
+let deleteData = async () => {
+  try {
+    await Tour.deleteMany()
+    console.log('data successfuly deleted!')
+  } catch(e) {
+    console.log(e)
+  }
+  process.exit()
+}
+
+let x = process.argv[2]
+if(x === '--import') importData()
+if(x === '--delete') deleteData()
